detector/requirements.txt
flask==3.0.3
pyyaml==6.0.1
requests==2.31.0
psutil==5.9.8